Replace manual addition of sites to Ansible
Manually adding sites to the Ansible config is a major PITA. Instead,
I realized we could just dump the ruby config to yaml and let Ansible
read it in like the rest of the config. Manipulating the variables in
Ruby is wayyyy easier than trying to do it all in nasty Jinja2 syntax.

I also replaced the function definitions with inline code to prevent the
need for making vconfig a global variable.